feat: split landing and workspace into routed pages
