--- src/app/app.html
@@ -1,14 +1,0 @@
-<div class="app" [class.app--landing]="!pdfDoc">
-  <ng-container *ngIf="!pdfDoc; else workspace">
-    <app-landing [languages]="languages" [selectedLanguage]="languageModel" [appName]="appName"
-      [version]="version" [currentYear]="currentYear" [appAuthor]="appAuthor"
-      [fileDropActive]="fileDropActive()" (languageChange)="onLanguageChange($event)"
-      (fileSelected)="onFileSelected($event)" (fileUploadKeydown)="onFileUploadKeydown($event)"
-      (fileDragOver)="onFileDragOver($event)" (fileDragLeave)="onFileDragLeave($event)"
-      (fileDrop)="onFileDrop($event)" (openFilePicker)="openPdfFilePicker()"></app-landing>
-  </ng-container>
-
-  <ng-template #workspace>
-    <app-workspace [vm]="vm"></app-workspace>
-  </ng-template>
-</div>
