+++ src/app/app.spec.ts
@@ -1,11 +1,11 @@
 import { TestBed } from '@angular/core/testing';
+import { RouterTestingModule } from '@angular/router/testing';
 import { App } from './app';
-import { TranslationService } from './i18n/translation.service';
 
 describe('App', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [App],
+      imports: [App, RouterTestingModule],
     }).compileComponents();
   });
 
@@ -15,11 +15,10 @@ describe('App', () => {
     expect(app).toBeTruthy();
   });
 
-  it('should render the localized title', () => {
+  it('should render the router outlet', () => {
     const fixture = TestBed.createComponent(App);
     fixture.detectChanges();
     const compiled = fixture.nativeElement as HTMLElement;
-    const translation = TestBed.inject(TranslationService);
-    expect(compiled.querySelector('.logo')?.textContent?.trim()).toBe(translation.translate('app.title'));
+    expect(compiled.querySelector('router-outlet')).not.toBeNull();
   });
 });
