+++ src/app/pages/workspace/workspace.page.html
@@ -1,0 +1,3 @@
+<div class="app">
+  <app-workspace [vm]="vm"></app-workspace>
+</div>
