+++ src/app/pages/workspace/workspace.page.scss
